Union de Todas los models

+++ Models/Aparment.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal PricePerNight { get; set; }
+        public double PricePerNight { get; set; }
         public string Address { get; set; }
         public string City { get; set; }
         public string Country { get; set; }
